refactor: extract thread size population
